refactor: SessionDetail.tsx からロジック・汎用コンポーネントを分離 (#156, #157, #158, #159)
- ストリームパーシングロジック・型定義を session/streamParsing.ts に分離
- ツールヘルパー関数を session/toolHelpers.ts に分離
- Toast コンポーネントを ui/Toast.tsx として汎用化
- useAutoScroll フックを hooks/useAutoScroll.ts に移動
