+++ frontend/src/components/ui/Toast.tsx
@@ -1,0 +1,15 @@
+import { CheckCircle2, AlertTriangle } from "lucide-react";
+
+export function Toast({ message, variant = "success" }: { message: string; variant?: "success" | "error" }) {
+  const styles = variant === "success"
+    ? "bg-green-600"
+    : "bg-red-600";
+  const Icon = variant === "success" ? CheckCircle2 : AlertTriangle;
+
+  return (
+    <div className={`fixed top-4 right-4 z-50 ${styles} text-white px-4 py-2 rounded shadow-lg text-sm flex items-center gap-2 animate-fade-in`}>
+      <Icon className="w-4 h-4" />
+      {message}
+    </div>
+  );
+}
