+++ frontend/src/hooks/useAutoScroll.ts
@@ -1,0 +1,24 @@
+import { useEffect, useRef } from "react";
+
+function isScrolledToBottom(el: HTMLElement, threshold = 50): boolean {
+  return el.scrollHeight - el.scrollTop - el.clientHeight < threshold;
+}
+
+export function useAutoScroll(dep: unknown) {
+  const ref = useRef<HTMLDivElement>(null);
+  const wasAtBottom = useRef(true);
+
+  useEffect(() => {
+    if (ref.current && wasAtBottom.current) {
+      ref.current.scrollTop = ref.current.scrollHeight;
+    }
+  }, [dep]);
+
+  const onScroll = () => {
+    if (ref.current) {
+      wasAtBottom.current = isScrolledToBottom(ref.current);
+    }
+  };
+
+  return { ref, onScroll };
+}
